test: Split fixtures into conftest.py
Split test fixtures into conftest.py to clean up and prepare for usage
of pytest_configure() hook. No functional changes involved.
